+++ config/credential.go
@@ -12,27 +12,6 @@ type CredentialConfig struct {
 	CredConfig
 }
 
-type KeyvaultCredentialConfig struct {
-	Name         string `yaml:"name,omitempty" validate:"required"`
-	TenantID     string `yaml:"tenantID,omitempty" validate:"required"`
-	ClientID     string `yaml:"clientID,omitempty" validate:"required"`
-	ClientSecret string `yaml:"clientSecret,omitempty" validate:"required"`
-}
-
-func (kvc KeyvaultCredentialConfig) GetName() string {
-	return kvc.Name
-}
-
-type CyberarkCredentialConfig struct {
-	Name         string `yaml:"name,omitempty" validate:"required"`
-	WorkloadID   string `yaml:"workloadID,omitempty" validate:"required"`
-	ApiKey       string `yaml:"apiKey,omitempty" validate:"required"`
-}
-
-func (kvc CyberarkCredentialConfig) GetName() string {
-	return kvc.Name
-}
-
 func (c *CredentialConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
 
 	var kvConfig KeyvaultCredentialConfig
